refactor: test layer와 production layer 분리
