+++ __tests__/Cars.test.js
@@ -1,7 +1,7 @@
 import Car from "../src/Models/Car";
 import { createCars } from "../src/Models/Cars";
 import { DuplicatedCarNameError } from "../src/Models/Cars/errors";
-import MoveStrategies from "../src/Models/MoveStrategy/MoveStrategies";
+import MoveStrategies from "../test/MoveStrategies";
 
 const { from, playOneRound, getRoundRecord } = createCars();
 describe("from() 테스트", () => {
