+++ test/EmptyStrategy.js
